Moved Workers to its own file
