+++ tests/ethereum/test_consensys_benchmark.py
@@ -16,7 +16,7 @@ class EthBenchmark(unittest.TestCase):
         self.worksp = self.mevm.workspace
 
     def tearDown(self):
-        self.mevm = None
+        del self.mevm
         shutil.rmtree(self.worksp)
 
     def _test(self, name, should_find, use_ctor_sym_arg=False):
